requirements.txt
pyyaml
networkx
jsonschema
